requirements.txt
Django
psycopg2-binary
pillow
ipython
djangorestframework
python-dotenv
faker
django-filter
djangorestframework-simplejwt
coverage